fix: rename type file
